# internal/handlers/apihandler/handler.go
// Copyright (c) Liam Stanley <liam@liam.sh>. All rights reserved. Use of
// this source code is governed by the MIT license that can be found in
// the LICENSE file.

package apihandler

import (
	_ "embed"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/httprate"
	"github.com/lrstanley/chix/v2"
	"github.com/lrstanley/geoip/internal/lookup"
)

type handler struct {
	lookupSvc *lookup.Service
	limiter   *httprate.RateLimiter
}

func New(lookupSvc *lookup.Service, limiter *httprate.RateLimiter) *handler { //nolint:revive
	return &handler{
		lookupSvc: lookupSvc,
		limiter:   limiter,
	}
}

func (h *handler) Route(r chi.Router) {
	r.Use(
		middleware.NoCache,
		middleware.GetHead,
	)

	// v1 API.
	r.With(h.limiter.Handler).Get("/{addr}", h.getLookupV1)
	r.With(httprate.Skip, h.limiter.Handler).Get("/ping", h.ping)

	// v2 API.
	r.Get("/v2/openapi.yaml", h.getOpenAPI)
	r.With(h.limiter.Handler).Get("/v2/lookup/{addr}", h.getLookupV2)
	r.With(h.limiter.Handler).Post("/v2/bulk", h.postBulkV2)
	r.With(h.limiter.Handler).Get("/v2/metadata", h.getMetadataV2)
	r.With(httprate.Skip, h.limiter.Handler).Get("/v2/ping", h.ping)
	r.NotFound(h.notFound)
}

func (h *handler) notFound(w http.ResponseWriter, r *http.Request) {
	chix.ErrorWithCode(w, r, http.StatusNotFound)
}

func (h *handler) ping(w http.ResponseWriter, r *http.Request) {
	chix.JSON(w, r, http.StatusOK, chix.M{"pong": true})
}

//go:embed openapi_v2.yaml
var openapiv2 string

func (h *handler) getOpenAPI(w http.ResponseWriter, _ *http.Request) {
	w.Header().Set("Content-Type", "application/yaml")
	_, _ = w.Write([]byte(openapiv2))
}